refactor(macos): simplify agent cursor implementation
