+++ native/macos/agent_cursor_motion.swift
